+++ src/components/home/landing.js
@@ -1,8 +1,10 @@
 import React from "react";
+import ScrollToAnchor from "../../utils/scrollToAnchor";
 
 const Landing = () => {
   return (
     <section className="grid md:grid-cols-2 gap-4 lg:mx-56 sm:mx-10 mx-5 min-h-screen">
+      <ScrollToAnchor />
       <div className="table">
         <div className="table-cell align-middle">
           <p className="subH">Krity Dangol</p>
@@ -13,7 +15,7 @@ const Landing = () => {
             laudantium accusantium alias vitae aliquam dolor laboriosam quo
             iusto!
           </p>
-          <a href="/#/categories" className="link">
+          <a href="/#/home#projects" className="link">
             View Projects
           </a>
           &nbsp;&nbsp;or&nbsp;&nbsp;